+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.0] - 2022-01-19
 ### Added
 - `bin` folder to put executables built with `pyinstaller`
+- Build script to ease creation of tags/releases
 - `test` folder and move test items into it
 - This changelog